// interactive-menu[mem]/utils/enums.ts
export enum WeaponModel {
    // Hand
    BrassKnuckle = 331,
    
    // Melee
    GolfClub = 333,
    Nightstick = 334,
    Knife = 335,
    Bat = 336,
    Shovel = 337,
    PoolStick = 338,
    Katana = 339,
    Chainsaw = 341,
    
    // Handguns
    Colt45 = 346,
    Silenced = 347,
    Deagle = 348,
    
    // Shotguns
    Shotgun = 349,
    SawedOff = 350,
    CombatShotgun = 351,
    
    // Sub-Machine Guns
    Uzi = 352,
    MP5 = 353,
    Tec9 = 372,
    
    // Assault Rifles
    AK47 = 355,
    M4 = 356,
    
    // Rifles
    Rifle = 357,
    Sniper = 358,
    
    // Heavy Weapons
    RocketLauncher = 359,
    RocketLauncherHS = 360,
    Flamethrower = 361,
    Minigun = 362,
    
    // Projectiles
    Grenade = 342,
    Teargas = 343,
    Molotov = 344,
    Satchel = 363,
    
    // Special 1
    Spraycan = 365,
    FireExtinguisher = 366,
    Camera = 367,
    
    // Gifts
    Dildo1 = 321,
    Dildo2 = 322,
    Vibrator = 323,
    Flowers = 325,
    Cane = 326,
    
    // Special 2
    Nightvision = 368,
    Infrared = 369,
    Parachute = 371,
    
    // Satchel Detonator
    SatchelDetonator = 364
}